Moved string literal logic to seperate module
